feat: Add updates to charts
