+++ src/components/Activity/components/components/index.tsx
@@ -1,0 +1,4 @@
+export { default as VizRating } from "./VizRating";
+export { default as VizHighestRated } from "./VizHighestRated";
+export { default as VizGenre } from "./VizGenre";
+export { default as VizReleased } from "./VizReleased";
