+++ src/components/Activity/components/index.tsx
@@ -1,4 +1,2 @@
-export { default as ChartAll } from "./ChartAll";
-export { default as ChartTop } from "./ChartTop";
-export { default as ChartYear } from "./ChartYear";
-export { ChartVizGenre, ChartVizRating, ChartVizReleased } from "./ChartViz";
+export { default as ActivityAll } from "./ActivityAll";
+export { default as ActivityYear } from "./ActivityYear";
