+++ src/components/Bookmarks/Bookmarks.tsx
@@ -12,12 +12,12 @@ import {
 import React, { useEffect } from "react";
 import { cache, useSWRInfinite } from "swr";
 import { useMDDispatch, useMDState } from "../../config/store";
-import { fuegoBookmarkGet } from "../../fuego/fuegoBookmarks";
 import type { MediaDiaryWithId } from "../../types/typesMedia";
 import type { UserFuegoValidated } from "../../types/typesUser";
 import { createPosterURL } from "../../utils/helpers";
 import LogoIcon from "../icons/LogoIcon";
 import MdLoader from "../md/MdLoader";
+import { fuegoBookmarkGet } from "./fuego";
 
 function Bookmarks({ user }: { user: UserFuegoValidated }): JSX.Element {
   const state = useMDState();
