+++ src/components/Bookmarks/fuego/index.tsx
@@ -1,0 +1,1 @@
+export { fuegoBookmarkGet } from "./fuegoBookmarks";
