+++ src/components/Day/fuego/index.tsx
@@ -1,0 +1,5 @@
+export {
+  fuegoBookmarkAddWithId,
+  fuegoBookmarkDelete,
+  fuegoBookmarkDeleteWithId,
+} from "./fuegoBookmarks";
